refactor(core): split storage manager actor in different files
